Inline SpillingJoinProcessor class in forked join operator
After the removal of spilling this class is fairly basic and in 1:1 relation to
LookupJoinOperator, so it can be easily inlined
