+++ core/trino-main/src/main/java/io/trino/operator/join/unspilled/LookupJoinOperator.java
@@ -13,6 +13,8 @@
  */
 package io.trino.operator.join.unspilled;
 
+import com.google.common.io.Closer;
+import com.google.common.util.concurrent.Futures;
 import com.google.common.util.concurrent.ListenableFuture;
 import io.trino.operator.OperatorInfo;
 import io.trino.operator.PageBuffer;
@@ -27,10 +29,13 @@ import io.trino.operator.join.unspilled.PageJoiner.PageJoinerFactory;
 import io.trino.spi.Page;
 import io.trino.spi.type.Type;
 
+import java.io.IOException;
 import java.util.List;
 import java.util.Optional;
 
+import static com.google.common.util.concurrent.MoreExecutors.directExecutor;
 import static io.trino.operator.WorkProcessor.flatten;
+import static java.util.Objects.requireNonNull;
 
 public class LookupJoinOperator
         implements AdapterWorkProcessorOperator
@@ -39,8 +44,13 @@ public class LookupJoinOperator
     private final boolean waitForBuild;
     private final PageBuffer pageBuffer;
     private final WorkProcessor<Page> pages;
-    private final SpillingJoinProcessor joinProcessor;
     private final JoinStatisticsCounter statisticsCounter;
+
+    private final Runnable afterClose;
+    private final PageJoiner sourcePagesJoiner;
+    private final WorkProcessor<Page> joinedSourcePages;
+
+    private boolean closed;
 
     LookupJoinOperator(
             List<Type> buildOutputTypes,
@@ -66,13 +76,11 @@ public class LookupJoinOperator
                         joinProbeFactory,
                         lookupSourceProvider,
                         statisticsCounter);
-        joinProcessor = new SpillingJoinProcessor(
-                afterClose,
-                waitForBuild,
-                lookupSourceFuture,
-                pageJoinerFactory,
-                sourcePages.orElse(pageBuffer.pages()));
-        pages = flatten(WorkProcessor.create(joinProcessor));
+        this.afterClose = requireNonNull(afterClose, "afterClose is null");
+        sourcePagesJoiner = pageJoinerFactory.getPageJoiner(lookupSourceFuture);
+        joinedSourcePages = sourcePages.orElse(pageBuffer.pages()).transform(sourcePagesJoiner);
+
+        pages = flatten(WorkProcessor.create(this::process));
     }
 
     @Override
@@ -105,9 +113,44 @@ public class LookupJoinOperator
         return pages;
     }
 
+    public WorkProcessor.ProcessState<WorkProcessor<Page>> process()
+    {
+        // wait for build side to be completed before fetching any probe data
+        // TODO: fix support for probe short-circuit: https://github.com/trinodb/trino/issues/3957
+        if (waitForBuild && !lookupSourceFuture.isDone()) {
+            return WorkProcessor.ProcessState.blocked(asVoid(lookupSourceFuture));
+        }
+
+        if (!joinedSourcePages.isFinished()) {
+            return WorkProcessor.ProcessState.ofResult(joinedSourcePages);
+        }
+
+        close();
+        return WorkProcessor.ProcessState.finished();
+    }
+
+    private static <T> ListenableFuture<Void> asVoid(ListenableFuture<T> future)
+    {
+        return Futures.transform(future, v -> null, directExecutor());
+    }
+
     @Override
     public void close()
     {
-        joinProcessor.close();
+        if (closed) {
+            return;
+        }
+        closed = true;
+
+        try (Closer closer = Closer.create()) {
+            // `afterClose` must be run last.
+            // Closer is documented to mimic try-with-resource, which implies close will happen in reverse order.
+            closer.register(afterClose::run);
+
+            closer.register(sourcePagesJoiner);
+        }
+        catch (IOException e) {
+            throw new RuntimeException(e);
+        }
     }
 }
